Updated as per review comments and tested again
